@@ -4,20 +4,20 @@
   "private": true,
   "dependencies": {
     "@ant-design/icons": "^4.6.4",
-    "@apollo/client": "^3.4.15",
-    "antd": "^4.16.13",
-    "graphql": "^15.6.0",
-    "i18next": "^20.6.1",
+    "@apollo/client": "^3.5.5",
+    "antd": "^4.17.2",
+    "graphql": "^16.0.1",
+    "i18next": "^21.5.4",
     "i18next-http-backend": "^1.3.1",
     "interactjs": "^1.10.11",
     "react": "^17.0.2",
     "react-apollo": "^3.1.5",
     "react-calendar-timeline": "^0.27.0",
     "react-dom": "^17.0.2",
-    "react-i18next": "^11.12.0",
-    "react-router-dom": "^5.3.0",
+    "react-i18next": "^11.14.3",
+    "react-router-dom": "^6.0.2",
     "react-scripts": "4.0.3",
-    "typescript": "^4.4.3"
+    "typescript": "^4.5.2"
   },
   "scripts": {
     "schema": "npx apollo client:download-schema --endpoint=http://localhost:8000/api schema.json",
@@ -47,10 +47,10 @@
   },
   "devDependencies": {
     "@types/graphql": "^14.5.0",
-    "@types/react-calendar-timeline": "^0.26.4",
-    "@types/react-dom": "^17.0.9",
-    "@types/react-router-dom": "^5.1.9",
-    "apollo": "^2.33.6"
+    "@types/react-calendar-timeline": "^0.26.5",
+    "@types/react-dom": "^17.0.11",
+    "@types/react-router-dom": "^5.3.2",
+    "apollo": "^2.33.9"
   },
   "proxy": "http://localhost:8000"
 }
