@@ -1,8 +1,9 @@
+import { useReactiveVar } from "@apollo/client"
 import Title from "antd/lib/typography/Title"
 import { pageTitle } from "../../../../cache"
 
 export const PageTitle = () => {
-
+  const title = useReactiveVar(pageTitle)
   return pageTitle() !== null ? (
     <Title
       level={ 3 }
@@ -10,7 +11,7 @@ export const PageTitle = () => {
         borderBottom: "1px solid #e1e1e1",
         paddingBottom: ".3em"
       } }>
-      { pageTitle() }
+      { title }
     </Title>
   ) : null
 }
